reposter/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# reposter/media_util.py: 1
Pillow == 7.0.0

# reposter/database/Security.py: 3,4,5,6
cryptography == 2.8

# reposter/instagram_util/InstaUtil.py: 1
# reposter/media_util.py: 95
git+git://github.com/pabloi09/igbot@master#egg=instabot

# reposter/daemons/insta_daemon.py: 2
lockfile == 0.12.2

# reposter/instagram_util/test_insta_util.py: 4
# reposter/media_util.py: 3
numpy == 1.19.4

# reposter/reddit_downloader/RedditDownloader.py: 1
praw == 7.1.0

# reposter/batch_processes/daily_scheduler.py: 1
python_crontab == 2.5.1

# reposter/daemons/insta_daemon.py: 1
python_daemon == 2.2.4

# reposter/reddit_downloader/RedditDownloader.py: 2
# reposter/twitter_util/VideoMedia.py: 5
requests == 2.22.0

# reposter/media_util.py: 2,56
scipy == 1.5.4

# reposter/setup.py: 1
setuptools == 45.2.0

# reposter/twitter_util/TwitterUtil.py: 1
tweepy == 3.9.0

# reposter/reddit_downloader/RedditDownloader.py: 7
youtube_dl == 2020.12.12

urllib3 == 1.25.8
